Redefine Entry class and introduce InstagramWebhook  for getting the webhhok request
Updated the `Entry` class to include `Messaging` and `Changes` properties.
Introduced the `InstagramWebhook` class with properties for `Object` and a list of `Entry` objects.
Added `ListPropertiesInTerminal` method to print details about the `Object` and its entries.
New classes `Value`, `User`, `Message`, and `Messaging` have been created to structure the data.